requirements.txt
factory-boy==3.3.0
pandas==2.1.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pytest==7.4.3